+++ app/page.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useEffect, useState } from "react";
 
 import Loading from "../components/Loading";
