+++ tailwind.config.js
@@ -6,6 +6,7 @@ module.exports = withMT({
   content: [
     "./pages/**/*.{js,ts,jsx,tsx}",
     "./components/**/*.{js,ts,jsx,tsx}",
+    "./app/**/*.{js,ts,jsx,tsx}",
     // material tailwind
     "./node_modules/@material-tailwind/react/**/*.{js,ts,jsx,tsx}",
   ],
